psuh: print the current branch
Demonstrate how to use wt-status to find information about the current
working tree's status.

The working tree status can reveal lots of interesting things about the
current context, for example, which branch the user is on, whether
changes are staged or not, and on which files. This can be extremely
handy to inform yourself of what the user is working on or likely to
care about. Luckily, cmd_status in builtin/commit.c is easy to read and
a good demonstration of how to use wt-status.h.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>