t: add psuh testing
Let's ensure that our core function is working as expected by checking
for the mention of a small horse.

In order to reduce test flakiness, here we test only for the mention of
the animal we seek. Thus, the test won't begin to fail when we change a
minor detail of the pony's appearance or the additional context details
it provides.  A stricter test could mean that any time a developer makes
a minor change to the printed text, they would need to modify the test;
this leads to developer toil and ill will towards you, the original
author. Best to keep your check concise.

The test_i18ngrep is a translation-aware grep that you should default to
using instead of normal grep when testing within Git.

In the unlikely event that users report a bug about git-psuh we should
add regression tests here accordingly during the course of the bugfix.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>