psuh: add usage text
Increase usability and discoverability by providing users some context
when they run `git psuh -h`.

This commit uses parse-options, which is described in
`Documentation/technical/api-parse-options.txt` and recommended for new
commands.  Furthermore, it fixes a test failure where `git psuh -h` is
not handled - parse-options handles this for you automatically.

parse-options is the recommended way to examine and handle command line
inputs. In fact, if a command does not plan to use parse-options, a
specific option must be provided in git.c when registering the command.
parse-options populates the option struct provided to it and modifies
the argv array by removing commands which were expected and provided;
unexpected commands will remain in argv, which allows us to still echo
all arguments the user provides to git psuh.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>