grep: provide pathspecs/patterns via file or stdin
Teach 'git grep' how to receive pathspecs via file, and add convenience
flags for receiving pathspecs or patterns via stdin.

In some cases, the list of pathspecs to 'git grep' is longer than the
command line allows, or long enough that a user becomes fatigued reading
the command.  In other cases, a user may wish to write a script or
oneliner which generates a list of files to 'git grep' against. In both
scenarios, a reasonable workaround is to provide the list of pathspecs
via a file or stdin.

Before this change, it was already possible to receive patterns via
stdin by using the argument '-f -'. Now that there are two arguments
which could both use stdin, help safeguard the user from doing so by
providing convenience '--stdin-patterns' and '--stdin-pathspecs' which are mutually exclusive.
(It is still possible to ask for '-f - --pathspecs-file -', though.)

Finally, in order to match '--pathspecs-file', give '-f' a long form
'--patterns-file' too.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>