ci: add null-checking around run.head_commit
In some cases it seems run.head_commit can be null while examining old
runs. When this happens, the current run aborts without evaluating the
commit.  Then, that same commit can never be retested - because it is
noted in the run history.

This commit doesn't fix the latter problem - that by reducing redundant
runs we leave ourselves at the mercy of GitHub Actions flakiness - but
does at least reduce the likelihood of failing a perfectly good commit
because some old run was malformed.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>