hook: allow stdin from strbuf
In some cases, such as "post-rewrite", it's easier to write to the
hook's stdin using a string and a pipe in memory than it is to find a
suitable temporary file.

Typically this is done like in `sequencer.c:run_rewrite_hook()` - set
child_process.in = -1 to indicate that you want a pipe created for you,
then call start_command() which creates the pipe, then fill in the pipe,
then call finish_command() which streams the pipe's contents to the
child process. However, the `run_processes_parallel()` API doesn't allow
us to touch the child_process after start_command() is called - so we
should set up the pipe for ourselves.

Because multiple children could run simultaneously, let's open and fill
a new pipe each time, so that children don't need to share a read fd.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>