hook API: "hook_state" member and track "active" explicitly
Amend code added in 96e7225b310 (hook: add 'run' subcommand,
2021-12-22) to be less confusing. We using the "hook_path" for two
purposes, to see what the path to the hook was, and as a marker in a
callback for the run_processes_parallel_tr2() machinery so that we
could abort in the pick_next_hook() callback.

Let's explicitly track an "active" variable instead ,and create a
"struct hook_state" to track the state of the hook execution.

This change makes subsequent changes less confusing, as we'll
eventually refactor away the "hook_path" member itself. Now that
change no longer needs to concern itself with the functionality of
this "active" tracking.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>