run-command: don't wait for stderr if more stdin is available
In run_processes_parallel(), we previously only needed to contend with
stderr to the child process - the child's stderr and stdout were
combined together and we forbade stdin. At that time, it made sense to
prevent ourselves from looping very tightly on stderr, so we set a 100ms
timeout to the poll() call on the child's stderr fd.

However, once processing of stdin was introduced, we saw a case where
some hooks exist only to consume lots of stdin, rarely write stderr/out,
and ultimately return an exit code which Git used to decide whether to
halt some operation. In these cases, because stderr's fd never presents
output for poll() to read, run_processes_parallel() waits the entire
100ms between each line of stdin provided to the hook.

Within pp_buffer_stdin(), we are already noticing whether the hook
callsite is finished writing all the stdin it wants to the hook; the
'feed_pipe' callback returns a "done" bit. Teach pp_buffer_stdin() to
return that "done" bit, and then use that bit in the
run_processes_parallel() IO loop to decide whether to wait a long time
for poll() on the child's stderr.

It could seem reasonable to write the entire stdin to the child before
doing any stderr processing; however, it could be possible for the
stderr pipe to fill up while we are busy writing stdin. For that reason,
we still want to do a quick (timeout=0) check on stderr between chunks of
stdin, and process anything that appeared in the pipe while we were
writing to stdin.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>