config: let feature.experimental imply gc.cruftPacks=true
We are interested in exploring whether gc.cruftPacks=true should become
the default value; to determine whether it is safe to do so, let's
encourage more users to try it out. Users who have set
feature.experimental=true have already volunteered to try new and
possibly-breaking config changes, so let's try this new default with
that set of users.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>