docs: propose more formalized decisionmaking
Note: This commit is a jumping-off point for discussion at Git Merge
2024. It is not ready to be merged as-is, or even necessarily
recommended by the author.

Some outstanding concerns with this WIP commit:
 * Generally, has a very bureaucratic feeling. How can we increase
   predictability without inflicting bureaucracy on our community?
 * Is prone to rushed "no" - if a reviewer doesn't have time within the
   stated timeline, the reviewer may say "no thanks", without developing
   the proposal.
 * Probably more..

Let's add more predictability to the decisionmaking process the project
uses. The goal of this change is to make it easier to tell *when* a
proposal will be approved or rejected, and to tell *how likely* each
outcome is as the proposal is being discussed. Both of these goals are
intended to make the discussion process less frustrating for the
proposer (and for the contributors who are participating).

The primary difference between our current process and this
process is the addition of explicit voting ("yes", "yes with changes",
"no") and a timeline ("please respond by YYYY-MM-DD").

Adding the explicit voting makes it easier to understand when there is
disagreement coming from a loud single party, and how to weigh it
against agreement from a quieter majority. Noticing these weights is
especially difficult when trying to gather the general feeling of
consensus in a long thread with some respondents more active than
others.

Adding the timeline makes the stakes easier to understand for the
proposer: rather than feeling stuck in limbo as the community discusses
the proposal over a long period of time, the proposer has a deadline
when they know they will be able to measure the community's approval.
This also provides a way for community members to decide how to
prioritize reviewing the proposal around other work they need to do:
deadlines are a great forcing function for time management.

It's still fairly informal - many methods of proposal are still valid,
and the timeline after proposal isn't rigorously set; specific deciders
or subsystem maintainers are not appointed or required to respond.

Change-Id: I7c652f4e7dfe3352943e37fe815e7a9e1fd79630
Signed-off-by: Emily Shaffer <emilyshaffer@google.com>